autotools: ensure aclocal runs successfully by passing -I.
autoreconf -fi can update the autotools files automatically, but if
aclocal doesn't know where project-local macros are it will delete them
and then the configure script is malformed and errors out.

This tells autoreconf to pass -I. to aclocal, that is, it should check
for stlport.m4 in the current directory and use it.